Converging all sqltoolsservice test to a single file since most of the code was redundant and hard to maintain.
